Add files via upload
This is just an update regarding the exponentiation of the parameters. 
We're still getting no convergence for most of the time. Success is random.
I have changed the RBF Kernel having l^2 instead of l in the denominator, since that ensures positivity. With this there is no need for further exponentiation. Still, by exponeniating l as well, we avoid the singularity for l = 0.
There is no need to exponeniate sigma, since it already appears squared and the factor is thus positive.